requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings 
python-dotenv
langchain
langchain-google-genai
langchain-community
pypdf
faiss-cpu
sentence-transformers 
requests 
streamlit
python-multipart